+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/kernel/util/GetterUtil_IW.java
@@ -1,0 +1,5 @@
+package com.liferay.portal.kernel.util;
+
+public class GetterUtil_IW {
+
+}